requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.2
tensorflow==2.13.1
shap==0.44.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pillow==10.1.0
matplotlib==3.7.1
seaborn==0.13.0
pydantic==2.5.0
python-dotenv==1.0.0